feat(scheduler): 적응형 질문 발행 — 전원 답변 기반, 자동교체 제거 (MG-16)
요구
시간 기준이 아니라 그룹 전원 답변 완료가 기준. 완료 직후 가장 가까운
KST 11시에 새 질문 발행. 자동교체 없음. 가입 즉시 첫 질문 발급.

변경
- scheduler.ts: 48h 자동 넘김 로직 제거. "최근 DQ 미완료면 무한정 대기,
  완료면 오늘자 신규 발급" 으로 단순화
- dailyQuestionCompletion.ts: tryFinalizeDailyQuestion의 date 이동 로직 제거.
  같은 날 두 개의 DQ가 필요한 룰(@@unique([familyId, date]) 충돌 방지) +
  완료 판정은 호출 측에서 런타임으로 수행
- QuestionService.getTodayQuestion / skipTodayQuestion: 48h carry-over →
  무기한 carry-over (이전 미완료 DQ가 며칠 지나도 그대로 표시)
- QuestionService.assignQuestionToFamily: private → public (가족 생성 시 호출)
- FamilyService.createFamily: 트랜잭션 후 즉시 첫 질문 발급. 실패해도
  가족 생성은 성공하도록 try/catch (다음 스케줄러 사이클에서 자동 재시도)

reminderScheduler.ts (KST 19시 미답변 알림 + 답변자 nudge): 검토 결과
ANSWER_REQUEST 7일 윈도우 / ANSWERER_NUDGE 4일 윈도우로 정상 동작.
기존 18 reminder 테스트 + 86 service 테스트 = 104/104 통과.

호환성
이미 finalize되어 date가 이동된 기존 DQ는 그대로 유지. 신규 DQ만
배정일 기준으로 저장되며 history 정렬은 date 컬럼으로 일관 유지.

Refs: MG-16
